changed spline functions from real to double precision

+++ code/src/mod_spline_functions.f08
@@ -9,7 +9,7 @@ contains
 
   subroutine quadratic_factors(r, rj_lo, rj, rj_hi, h_quadratic)
     double precision, intent(in)  ::  r, rj_lo, rj, rj_hi
-    real, intent(out)             ::  h_quadratic(4)
+    double precision, intent(out) ::  h_quadratic(4)
 
     h_quadratic(1) = 4.0d0 * (r - rj_lo) * (rj - r) / (rj - rj_lo)**2
     h_quadratic(2) = 0.0d0
@@ -21,7 +21,7 @@ contains
 
   subroutine quadratic_factors_deriv(r, rj_lo, rj, rj_hi, dh_quadratic_dr)
     double precision, intent(in)  ::  r, rj_lo, rj, rj_hi
-    real, intent(out)             ::  dh_quadratic_dr(4)
+    double precision, intent(out) ::  dh_quadratic_dr(4)
 
     dh_quadratic_dr(1) = 4.0d0 * (-2.0d0*r + rj + rj_lo) / (rj - rj_lo)**2
     dh_quadratic_dr(2) = 0.0d0
@@ -33,7 +33,7 @@ contains
 
   subroutine cubic_factors(r, rj_lo, rj, rj_hi, h_cubic)
     double precision, intent(in)  :: r, rj_lo, rj, rj_hi
-    real, intent(out)             :: h_cubic(4)
+    double precision, intent(out) :: h_cubic(4)
 
     h_cubic(1) =  3.0d0 * ( (r - rj_lo) / (rj - rj_lo) )**2 &
                  -2.0d0 * ( (r - rj_lo) / (rj - rj_lo) )**3
@@ -47,7 +47,7 @@ contains
 
   subroutine cubic_factors_deriv(r, rj_lo, rj, rj_hi, dh_cubic_dr)
     double precision, intent(in)  :: r, rj_lo, rj, rj_hi
-    real, intent(out)             :: dh_cubic_dr(4)
+    double precision, intent(out) :: dh_cubic_dr(4)
 
     dh_cubic_dr(1) =  6.0d0 * (r - rj_lo) / (rj - rj_lo)**2 &
                      -6.0d0 * (r - rj_lo)**2 / (rj - rj_lo)**3
